refactor(vite-plugin-mercury): extract test utilities to shared file
- Create src/test-utils.ts with renderHtml and createDOMFromHTML functions
- Refactor remark-inline-code.test.ts to use shared test utilities
- Add test-utils.ts to package.json files exclusion list
- Remove duplicate test helper code for better maintainability
